Migrate Tag definition to within TagManager.swift
