--- bacon/model/Tags/Tag.swift
@@ -1,81 +1,0 @@
-//
-//  Tag.swift
-//  bacon
-//
-//  Created by Fabian Terh on 10/4/19.
-//  Copyright © 2019 nus.CS3217. All rights reserved.
-//
-
-import Foundation
-
-/// Tag ADT to represent a transaction tag.
-/// A Tag has a value, and an optional parent value.
-/// A Tag must be either a parent or child Tag.
-struct Tag: Codable, Comparable, Hashable {
-
-    let value: String
-    let parent: String?
-
-    /// Initializes a parent Tag.
-    init(_ value: String) {
-        self.init(value, parent: nil)
-    }
-
-    /// Initializes a child Tag.
-    init(_ value: String, parent: String?) {
-        self.value = value
-        self.parent = parent
-    }
-
-    /// Convenience computed property to represent whether a Tag is a child Tag.
-    var isChild: Bool {
-        return parent != nil
-    }
-
-    /// Convenience computed property to represent whether a Tag is a parent Tag.
-    var isParent: Bool {
-        return !isChild
-    }
-
-    // Generally, we should only compare by the `value` property, since it's
-    // not meaningful to compare children Tags (of different parents), or a child Tag with a parent Tag.
-    static func < (lhs: Tag, rhs: Tag) -> Bool {
-        // 4 possible scenarios:
-        // 1) child vs child,
-        // 2) parent vs parent,
-        // 3) child vs parent,
-        // 4) parent vs child
-
-        if lhs.isChild && rhs.isChild {
-            // Compare their `value` properties first, then compare their `parent` properties
-            if lhs.value != rhs.value {
-                return lhs.value < rhs.value
-            }
-
-            // 2 sub-scenarios: (1) same parent, (2) different parents
-            guard let lhsParent = lhs.parent else {
-                fatalError("This should never happen")
-            }
-            guard let rhsParent = rhs.parent else {
-                fatalError("This should never happen")
-            }
-            assert(lhsParent != rhsParent)
-            return lhsParent < rhsParent
-        } else if lhs.isParent && rhs.isParent {
-            return lhs.value < rhs.value
-        } else if lhs.isChild && rhs.isParent {
-            // Compare their `value` properties first.
-            // If equal, a parent Tag should come before a child Tag.
-            if lhs.value != rhs.value {
-                return lhs.value < rhs.value
-            }
-            return false
-        } else { // lhs.isParent && rhs.isChild
-            if lhs.value != rhs.value {
-                return lhs.value < rhs.value
-            }
-            return true
-        }
-    }
-
-}
